Change default package level to 'lev1' in cli args
The default value for the CLI argument related to the level of included packages has been modified from 'lev2' to 'lev1'. This alteration has been made to limit the default level of package details shown to the users for better legibility. The change can be found in 'cli.rs' file.